move mocked_drop_for_seqrepo into dedicated pytest plugin
